--- java/src/Tile.java
@@ -1,7 +1,0 @@
-
-public class Tile
-{
-    public int X;
-    public int Y;
-    public char Symbol;
-}
