+++ java/src/main/java/Tile.java
@@ -1,0 +1,7 @@
+
+public class Tile
+{
+    public int X;
+    public int Y;
+    public char Symbol;
+}
