+++ java/src/test/java/GameTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class Game_Should {
+public class GameTest {
     private Game game;
 
     @BeforeEach
